refactor membership invitation token handler to another file
