+++ src/components/AccountCreatedPage.tsx
@@ -1,114 +1,20 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import { useSourceApp, SourceAppLogo } from './SourceApp'
 import { Button, Link, Grid } from '@mui/material'
 import { AppContextConsumer } from 'AppContext'
-import {
-  SynapseClient,
-  SynapseComponents,
-  Typography,
-} from 'synapse-react-client'
+import { Typography } from 'synapse-react-client'
 import { Link as RouterLink } from 'react-router-dom'
 import SourceAppConfigs from './SourceAppConfigs'
 import { LeftRightPanel } from './LeftRightPanel'
 import { sage } from 'configs/sagebionetworks'
-import { isMembershipInvtnSignedToken } from 'synapse-react-client/dist/utils/synapseTypes/SignedToken/MembershipInvtnSignedToken'
-import { useSynapseContext } from 'synapse-react-client/dist/utils/SynapseContext'
-import { displayToast } from 'synapse-react-client/dist/containers/ToastMessage'
-import { SignedTokenInterface } from 'synapse-react-client/dist/utils/synapseTypes/SignedToken/SignedTokenInterface'
-
+import { MembershipInvitationTokenHandler } from './MembershipInvitationTokenHandler'
 export type AccountCreatedPageProps = {}
 
 export const AccountCreatedPage = (props: AccountCreatedPageProps) => {
-  const { accessToken } = useSynapseContext()
-  const [signedToken, setSignedToken] = useState<
-    SignedTokenInterface | undefined
-  >()
-  const localStorageSignedToken = localStorage.getItem('signedToken')
-  useEffect(() => {
-    if (localStorageSignedToken) {
-      setSignedToken(
-        JSON.parse(
-          SynapseComponents.hex2ascii(localStorageSignedToken),
-        ) as SignedTokenInterface,
-      )
-    }
-  }, [localStorageSignedToken])
-
-  // If we have a MembershipInvtnSignedToken in the local storage, then process and clear it!
-  useEffect(() => {
-    const fetchData = async () => {
-      if (
-        accessToken &&
-        signedToken &&
-        isMembershipInvtnSignedToken(signedToken)
-      ) {
-        let membershipInvitation = await SynapseClient.getMembershipInvitation(
-          signedToken,
-        )
-        if (!membershipInvitation.inviteeId) {
-          // email is filled in, we must first bind the invitation
-          const inviteeSignedToken =
-            await SynapseClient.getInviteeVerificationSignedToken(
-              membershipInvitation.id,
-              accessToken,
-            )
-          try {
-            // attempt to bind the membership invite to new account
-            await SynapseClient.bindInvitationToAuthenticatedUser(
-              inviteeSignedToken,
-              membershipInvitation.id,
-              accessToken,
-            )
-            membershipInvitation = await SynapseClient.getMembershipInvitation(
-              signedToken,
-            )
-          } catch (error) {
-            if (error?.status === 403) {
-              displayToast(
-                `Couldn't join the team. This invitation was sent to an email address not associated to the current user. ${membershipInvitation.inviteeEmail} Please add this email in your Account Settings, or log in with the correct account before accepting the invitation.`,
-                'warning',
-                {
-                  primaryButtonConfig: {
-                    text: 'Account Settings',
-                    href: '/authenticated/myaccount',
-                  },
-                },
-              )
-            } else {
-              throw error
-            }
-          }
-        }
-
-        // returns undefined (if inviteeId is null)
-        if (membershipInvitation.inviteeId) {
-          await SynapseClient.addTeamMemberAsAuthenticatedUserOrAdmin(
-            membershipInvitation.teamId,
-            membershipInvitation.inviteeId,
-            accessToken,
-          )
-          // clear token from local storage, and show success UI
-          localStorage.removeItem('signedToken')
-          setSignedToken(undefined)
-          displayToast(`Successfully joined the team.`)
-        }
-      }
-    }
-
-    if (
-      accessToken &&
-      signedToken &&
-      isMembershipInvtnSignedToken(signedToken)
-    ) {
-      fetchData().catch(err => {
-        displayToast(err.reason, 'danger')
-      })
-    }
-  }, [signedToken])
-
   const sourceApp = useSourceApp()
   return (
     <>
+      <MembershipInvitationTokenHandler />
       <AppContextConsumer>
         {appContext => (
           <LeftRightPanel
